+++ addOns/wappalyzer/CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## [21.40.0] - 2024-08-28
 ### Changed
 - Updated with enthec upstream icon and pattern changes.
 - The add-on now has an options screen to allow users to select Quick/Exhaustive mode used by the passive scanner (part of Issue 8361).
@@ -381,6 +381,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - First version
 
 
+[21.40.0]: https://github.com/zaproxy/zap-extensions/releases/wappalyzer-v21.40.0
 [21.39.0]: https://github.com/zaproxy/zap-extensions/releases/wappalyzer-v21.39.0
 [21.38.0]: https://github.com/zaproxy/zap-extensions/releases/wappalyzer-v21.38.0
 [21.37.0]: https://github.com/zaproxy/zap-extensions/releases/wappalyzer-v21.37.0